packages/shared/pyproject.toml
[project]
name = "shared"
version = "0.1.0"
description = "Shared types, prompts, and A2A helpers for AgentMesh"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.11",
    "a2a-sdk[http-server]==0.3.26",
    "langchain>=0.3",
    "langchain-groq>=0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/shared"]
